factor out CheckpointWorker, tests and util structs into modules
