+++ rust/kafka-deduplicator/src/checkpoint/mod.rs
@@ -6,6 +6,7 @@ pub mod metadata;
 pub mod s3_client;
 pub mod s3_uploader;
 pub mod uploader;
+pub mod worker;
 
 pub use client::CheckpointClient;
 pub use config::CheckpointConfig;
@@ -15,3 +16,7 @@ pub use metadata::{CheckpointFile, CheckpointInfo, CheckpointMetadata, Checkpoin
 pub use s3_client::S3CheckpointClient;
 pub use s3_uploader::S3Uploader;
 pub use uploader::CheckpointUploader;
+pub use worker::{
+    CheckpointMode, CheckpointPath, CheckpointWorker, CHECKPOINT_PARTITION_PREFIX,
+    CHECKPOINT_TOPIC_PREFIX,
+};
